refactor: misc. moving files around
